move keras derived class to keras_utils
